geotiff tests: consolidate attrs-contract cluster (#2397)
Folds the four ``test_attrs_contract_*_1984.py`` files into one
parametrised ``attrs/test_contract.py`` under the new ``attrs/``
package. Tests are organised by contract dimension (canonical /
aliases / passthrough / version) with descriptive parametrise IDs
(``canonical[GDAL_NODATA]``, ``alias[nodatavals->nodata]``, etc.)
rather than per-issue filenames.

- 4 source files deleted, 1 new file added (-3 net).
- 40 test cases retained; coverage parity verified in
  ``CLUSTER_AUDIT_PR5.md`` (audit file deleted before merge).
- Release-gate checklist (release_gate_geotiff.rst) updated to point
  at the new path; the cited-files audit gate stays green.
- Markers come from ``_helpers/markers.py`` (no per-file gpu probe).
- Tests-only restructure; no source changes.

Part of epic #2390 (GeoTIFF test consolidation, PR 5 of 11).
